refactor(group-quiz): extract ResultCard into colocated component
Move ResultCard plus its exclusive deps (ResultTone, RESULT_TONES,
TrailerData, ResultCardProps) verbatim into _components/result-card.tsx
— none were used by the page. Drops the now-unused MusicPreview and
PlayCircle imports. Behavior-identical; page.tsx 1637->1314 LOC.
